+++ filters/GeoreferenceFilter.cpp
@@ -1,42 +1,41 @@
 /******************************************************************************
-* Copyright (c) 2023, Guilhem Villemin (guilhem.villemin@altametris.com)
-*
-* All rights reserved.
-*
-* Redistribution and use in source and binary forms, with or without
-* modification, are permitted provided that the following
-* conditions are met:
-*
-*     * Redistributions of source code must retain the above copyright
-*       notice, this list of conditions and the following disclaimer.
-*     * Redistributions in binary form must reproduce the above copyright
-*       notice, this list of conditions and the following disclaimer in
-*       the documentation and/or other materials provided
-*       with the distribution.
-*     * Neither the name of Hobu, Inc. or Flaxen Consulting LLC nor the
-*       names of its contributors may be used to endorse or promote
-*       products derived from this software without specific prior
-*       written permission.
-*
-* THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-* "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-* LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-* FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-* COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-* INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-* BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-* OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
-* AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-* OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
-* OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
-* OF SUCH DAMAGE.
-****************************************************************************/
+ * Copyright (c) 2023, Guilhem Villemin (guilhem.villemin@altametris.com)
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following
+ * conditions are met:
+ *
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in
+ *       the documentation and/or other materials provided
+ *       with the distribution.
+ *     * Neither the name of Hobu, Inc. or Flaxen Consulting LLC nor the
+ *       names of its contributors may be used to endorse or promote
+ *       products derived from this software without specific prior
+ *       written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
+ * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
+ * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
+ * OF SUCH DAMAGE.
+ ****************************************************************************/
 
+#include "private/georeference/LocalCartesian.hpp"
+#include "private/georeference/Trajectory.hpp"
+#include "private/georeference/Utils.hpp"
 #include <pdal/filters/TransformationFilter.hpp>
-#include "private/LocalCartesian.hpp"
-#include "private/Trajectory.hpp"
-#include "private/Utils.hpp"
-
 
 #include "GeoreferenceFilter.hpp"
 
@@ -44,10 +43,11 @@
 
 namespace pdal
 {
-static PluginInfo const s_info{"filters.georeference", "Georeferencing filter",
-                               "http://link/to/documentation"};
+static PluginInfo const s_info{
+    "filters.georeference", "Georeferencing filter",
+    "http://pdal.io/stages/filters.georeference.html"};
 
-CREATE_SHARED_STAGE(GeoreferenceFilter, s_info)
+CREATE_STATIC_STAGE(GeoreferenceFilter, s_info)
 
 using DimId = Dimension::Id;
 
@@ -70,9 +70,9 @@ public:
 };
 
 GeoreferenceFilter::GeoreferenceFilter()
-    : Filter(), Streamable(), m_config(nullptr), m_reverse(false),
-      m_matrix(new TransformationFilter::Transform), m_timeOffset(0.0),
-      m_localCartesian(new LocalCartesian(0.0, 0.0, 0.0))
+    : Filter(), Streamable(), m_matrix(new TransformationFilter::Transform),
+      m_config(nullptr), m_localCartesian(new LocalCartesian(0.0, 0.0, 0.0)),
+      m_trajectory(""), m_scan2imu(""), m_timeOffset(0.0), m_reverse(false)
 {
 }
 GeoreferenceFilter::~GeoreferenceFilter() {}
