+++ filters/GeoreferenceFilter.hpp
