--- plugins/georeference/filters/private/Trajectory.hpp
@@ -1,66 +1,0 @@
-/******************************************************************************
-* Copyright (c) 2023, Guilhem Villemin (guilhem.villemin@altametris.com)
-*
-* All rights reserved.
-*
-* Redistribution and use in source and binary forms, with or without
-* modification, are permitted provided that the following
-* conditions are met:
-*
-*     * Redistributions of source code must retain the above copyright
-*       notice, this list of conditions and the following disclaimer.
-*     * Redistributions in binary form must reproduce the above copyright
-*       notice, this list of conditions and the following disclaimer in
-*       the documentation and/or other materials provided
-*       with the distribution.
-*     * Neither the name of Hobu, Inc. or Flaxen Consulting LLC nor the
-*       names of its contributors may be used to endorse or promote
-*       products derived from this software without specific prior
-*       written permission.
-*
-* THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-* "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-* LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-* FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-* COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-* INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-* BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-* OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
-* AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-* OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
-* OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY
-* OF SUCH DAMAGE.
-****************************************************************************/
-
-#pragma once
-#include <pdal/PointTable.hpp>
-#include <pdal/PointView.hpp>
-#include <pdal/pdal_types.hpp>
-
-#include <filesystem>
-#include <list>
-#include <map>
-#include <string>
-#include <utility>
-
-namespace pdal
-{
-struct TrajPoint;
-class Trajectory
-{
-  pdal::PointViewPtr m_pointView;
-  pdal::PointTable m_table;
-  pdal::PointViewSet m_set;
-
-public:
-  Trajectory(const std::filesystem::path& trajFile);
-
-  bool getTrajPoint(double time, TrajPoint& trajPoint) const;
-};
-
-struct TrajPoint
-{
-  double roll, pitch, azimuth, wanderAngle, x, y, z, time;
-};
-
-};  // namespace pdal
